flattenTerms2 verifying working
