+++ src/javasources/KTool/src/org/kframework/compile/transformers/FlattenTerms.java
@@ -77,6 +77,11 @@ public class FlattenTerms extends CopyOnWriteTransformer {
                 node.setChild(child);
                 node.setLabel(label);
             }
+            if(!(node.getChild() instanceof KList)){
+                KList lok = new KList(label.getLocation(), label.getFilename());
+                lok.getContents().add(node.getChild());
+                node.setChild(lok);
+            }
             return node;
         }
